getRoomMessages
