+++ apps/meteor/app/livechat/server/lib/LivechatTyped.ts
@@ -7,7 +7,6 @@ import type {
 	IOmnichannelRoom,
 	IOmnichannelRoomClosingInfo,
 	IUser,
-	MessageTypesValues,
 	ILivechatVisitor,
 	SelectedAgent,
 	ILivechatAgent,
@@ -942,26 +941,6 @@ class LivechatClass {
 				type: 'agentStatus',
 				status,
 			});
-		});
-	}
-
-	async getRoomMessages({ rid }: { rid: string }) {
-		const room = await Rooms.findOneById(rid, { projection: { t: 1 } });
-		if (room?.t !== 'l') {
-			throw new Meteor.Error('invalid-room');
-		}
-
-		const ignoredMessageTypes: MessageTypesValues[] = [
-			'livechat_navigation_history',
-			'livechat_transcript_history',
-			'command',
-			'livechat-close',
-			'livechat-started',
-			'livechat_video_call',
-		];
-
-		return Messages.findVisibleByRoomIdNotContainingTypes(rid, ignoredMessageTypes, {
-			sort: { ts: 1 },
 		});
 	}
